requirements.txt
algorithmia>=1.0.0,<2.0
PyYAML==5.3.1
scikit-learn==0.22
matplotlib==3.01
seaborn==0.11.0
xgboost==1.2.1
psycopg2

Flask
